server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai
pymongo[srv]==1.0.0
gunicorn==21.2.0
